feat(*): refactor bad code and initialize clean react instance
