+++ package.json
@@ -8,6 +8,7 @@
     "build": "tsc -noEmit -skipLibCheck && node esbuild.config.mjs production",
     "prettier": "prettier --write 'src/**/*.+(ts|tsx|json|html|css)' 'styles.css'",
     "eslint": "eslint . --ext .ts,.tsx --fix",
+    "lint": "eslint src --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
     "version": "node version-bump.mjs && git add manifest.json versions.json"
   },
   "author": "Pash Shocky",
@@ -29,12 +30,13 @@
     "builtin-modules": "3.3.0",
     "esbuild": "^0.15.7",
     "eslint": "^8.51.0",
+    "eslint-plugin-import": "^2.28.1",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.4.1",
     "obsidian": "latest",
     "prettier": "^2.6.2",
     "tslib": "2.4.0",
-    "typescript": "^5.1.3"
+    "typescript": "4.4.4"
   },
   "dependencies": {
     "react": "^18.2.0",
